+++ spec/onlyoffice_testrail_wrapper/testrail2_run_spec.rb
@@ -22,4 +22,17 @@ describe OnlyofficeTestrailWrapper::Testrail2, '#run' do
     run.delete
     expect(project.get_run_by_name(run_name)).to be_nil
   end
+
+  describe '#is_completed' do
+    it 'By default run is not closed' do
+      run = project.create_new_run(run_name, suite.id)
+      expect(run.is_completed).to be_falsey
+    end
+
+    it 'Run can be closed (completed)' do
+      run = project.create_new_run(run_name, suite.id)
+      run.close
+      expect(project.get_run_by_name(run_name)).to be_truthy
+    end
+  end
 end